COLMAPtry/implementation/requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
PyYAML>=5.4.0
colorlog>=6.0.0
tqdm>=4.60.0
pillow>=8.0.0
plyfile>=0.7.4
trimesh>=3.9.0
pytest>=7.0.0

